Condense types to a single exportable type folder
