+++ package.json
@@ -13,8 +13,7 @@
     "not IE 11"
   ],
   "scripts": {
-    "build": "wp-scripts build --webpack-copy-php --webpack-src-dir=blocks",
-    "create-block": "npx @alleyinteractive/create-block -n wp-page-cache-control",
+    "build": "wp-scripts build",
     "precheck-types": "check-node-version --package",
     "check-types": "tsc",
     "lint:fix": "eslint --ext .jsx --ext .js --ext .ts --ext .tsx . --fix",
@@ -29,80 +28,20 @@
     "prestylelint": "check-node-version --package",
     "pretest:watch": "check-node-version --package",
     "pretest": "check-node-version --package",
-    "start": "wp-scripts start --webpack-copy-php --webpack-src-dir=blocks",
-    "start:hot": "wp-scripts start --webpack-copy-php --webpack-src-dir=blocks --hot",
+    "start": "wp-scripts start",
+    "start:hot": "wp-scripts start --hot",
     "stylelint:fix": "stylelint --fix \"**/*.scss\"",
     "stylelint": "stylelint \"**/*.scss\" --allow-empty-input",
     "test:watch": "jest --watch",
     "test": "jest --passWithNoTests"
   },
-  "dependencies": {
-    "@alleyinteractive/block-editor-tools": "^0.3.2",
-    "@types/jest": "^29.5.2",
-    "@types/wordpress__block-editor": "^11.5.0",
-    "@wordpress/a11y": "^3.30.0",
-    "@wordpress/annotations": "^2.30.0",
-    "@wordpress/api-fetch": "^6.27.0",
-    "@wordpress/autop": "^3.30.0",
-    "@wordpress/blob": "^3.30.0",
-    "@wordpress/block-directory": "^4.7.0",
-    "@wordpress/block-editor": "^11.7.0",
-    "@wordpress/block-library": "^8.7.0",
-    "@wordpress/block-serialization-default-parser": "^4.30.0",
-    "@wordpress/blocks": "^12.7.0",
-    "@wordpress/components": "^23.7.0",
-    "@wordpress/compose": "^6.7.0",
-    "@wordpress/core-data": "^6.7.0",
-    "@wordpress/customize-widgets": "^4.7.0",
-    "@wordpress/data": "^9.0.0",
-    "@wordpress/data-controls": "^2.30.0",
-    "@wordpress/date": "^4.30.0",
-    "@wordpress/deprecated": "^3.30.0",
-    "@wordpress/dom": "^3.30.0",
-    "@wordpress/dom-ready": "^3.30.0",
-    "@wordpress/edit-post": "^7.7.0",
-    "@wordpress/edit-site": "^5.7.0",
-    "@wordpress/edit-widgets": "^5.7.0",
-    "@wordpress/element": "^5.7.0",
-    "@wordpress/escape-html": "^2.30.0",
-    "@wordpress/format-library": "^4.7.0",
-    "@wordpress/hooks": "^3.30.0",
-    "@wordpress/html-entities": "^3.30.0",
-    "@wordpress/i18n": "^4.30.0",
-    "@wordpress/icons": "^9.21.0",
-    "@wordpress/interface": "^5.7.0",
-    "@wordpress/is-shallow-equal": "^4.30.0",
-    "@wordpress/keyboard-shortcuts": "^4.7.0",
-    "@wordpress/keycodes": "^3.30.0",
-    "@wordpress/list-reusable-blocks": "^4.7.0",
-    "@wordpress/media-utils": "^4.21.0",
-    "@wordpress/notices": "^3.30.0",
-    "@wordpress/nux": "^6.0.0",
-    "@wordpress/plugins": "^5.7.0",
-    "@wordpress/primitives": "^3.28.0",
-    "@wordpress/priority-queue": "^2.30.0",
-    "@wordpress/redux-routine": "^4.30.0",
-    "@wordpress/reusable-blocks": "^4.7.0",
-    "@wordpress/rich-text": "^6.7.0",
-    "@wordpress/scripts": "^26.1.0",
-    "@wordpress/server-side-render": "^4.7.0",
-    "@wordpress/shortcode": "^3.30.0",
-    "@wordpress/token-list": "^2.30.0",
-    "@wordpress/url": "^3.31.0",
-    "@wordpress/viewport": "^5.7.0",
-    "@wordpress/warning": "^2.30.0",
-    "@wordpress/widgets": "^3.7.0",
-    "@wordpress/wordcount": "^3.30.0",
-    "dompurify": "^3.0.4",
-    "prop-types": "^15.8.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-  },
   "devDependencies": {
     "@alleyinteractive/eslint-config": "^0.1.0",
     "@alleyinteractive/stylelint-config": "^0.0.2",
     "@babel/preset-env": "^7.22.7",
+    "@types/jest": "^29.5.2",
     "@wordpress/babel-preset-default": "^7.14.0",
+    "@wordpress/scripts": "^26.1.0",
     "babel-jest": "^29.6.1",
     "check-node-version": "^4.2.1",
     "clean-webpack-plugin": "^4.0.0",
